pyproject.toml
[project]
name = "bgp-mining"
version = "0.1.0"
description = ""
authors = [
    {name = "Caio Fontes",email = "caio.fontes.gondin@hotmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "rq (>=2.7.0,<3.0.0)",
    "apscheduler (>=3.11.2,<4.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "ipaddress (>=1.0.23,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "streamlit (>=1.45.0,<2.0.0)",
    "pandas (>=2.2.0,<3.0.0)",
    "streamlit-autorefresh (>=1.0.1,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "ruff (>=0.15.6,<0.16.0)",
    "taskipy (>=1.14.1,<2.0.0) ; python_version < '4.0'",
    "typos (>=1.44.0,<2.0.0)"
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.taskipy.tasks]
# ─── Qualidade de código ──────────────────────────────────────────────────────
lint   = "ruff check ."
format = "ruff format ."
spell  = "typos ."
test   = "pytest -v"
check  = "task lint && task spell && task test"

# ─── Firewall (necessário para Docker acessar internet via NAT) ───────────────
firewall-fix = "sudo sed -i 's/DEFAULT_FORWARD_POLICY=\"DROP\"/DEFAULT_FORWARD_POLICY=\"ACCEPT\"/' /etc/default/ufw && sudo ufw reload"

# ─── Docker prod (tudo containerizado) ───────────────────────────────────────
up     = "DOCKER_BUILDKIT=0 docker compose -f docker-compose.prod.yml up -d"
down   = "docker compose -f docker-compose.prod.yml down"
logs   = "docker compose -f docker-compose.prod.yml logs -f"
build  = "DOCKER_BUILDKIT=0 docker compose -f docker-compose.prod.yml build"
reset  = "docker compose -f docker-compose.prod.yml down -v && DOCKER_BUILDKIT=0 docker compose -f docker-compose.prod.yml up -d --build"

# ─── Docker dev (só infra — serviços rodam localmente) ────────────────────────
dev-up   = "docker compose -f docker-compose.dev.yml up -d"
dev-down = "docker compose -f docker-compose.dev.yml down"
dev-logs = "docker compose -f docker-compose.dev.yml logs -f"

# ─── Coletores (execução direta, sem Docker) ──────────────────────────────────
collect-ripe  = "python collectors/ripe/collector.py"
collect-bgp   = "python collectors/bgptools/collector.py"
collect-all   = "task collect-ripe && task collect-bgp"

# ─── Banco de dados ───────────────────────────────────────────────────────────
db-init  = "docker compose -f docker-compose.prod.yml exec postgres psql -U bgp -d bgp_mining -f /docker-entrypoint-initdb.d/init.sql"
db-shell = "docker compose -f docker-compose.prod.yml exec postgres psql -U bgp -d bgp_mining"

# ─── Worker / scheduler / dashboard (execução direta, sem Docker) ─────────────
worker    = "python worker/worker.py"
scheduler = "python scheduler/scheduler.py"
dashboard = "streamlit run dashboard/app.py"

# ─── Pesquisa de mitigadores ─────────────────────────────────────────────────
find-mitigators     = "python scripts/find_mitigators.py"
find-mitigators-sql = "python scripts/find_mitigators.py --output sql"